Use a canonical hostname for staging
On the server side the DNS hostname, domainpart of the JID and
server configured host should all match. If they differ, particularly
if the JID domainpart stored in the database differs from the server
configured host, then Bad Things can happen. This leads to subtle
bugs that are very difficult to track down.

This change reverts back to 'staging.dev.tinyrobot.com' as the
canonical hostname for staging, which means the configured host, DNS
and database should all match.